MSTM3Runner

MSTM3Runner.read_tmatrix — Method

read_tmatrix(filename)

Read the MSTM3 format T-Matrix.

Note that in MSTM3, TM = 1 and TE = 2, which is the opposite of the definition in TransitionMatrices.jl.

MSTM3Runner.run_mstm — Method

run_mstm(cfg; keep, mstm_exe_name, mstm_command)

Use the given configuration to run MSTM v3.

If keep = true, the working directory will not be removed after the run.

mstm_exe_name specifies the name or path of your compiled MSTM v3 executable.

MSTM3Runner.write_tmatrix — Method

write_tmatrix(filename, 𝐓, shape)

Write the T-Matrix to filename in the format expected by MSTM3.

Note that in MSTM3, TM = 1 and TE = 2, which is the opposite of the definition in TransitionMatrices.jl.